moving customer utility functions unrelated to rendering out of the index route file
